refactor: decouple history report types
